requirements.txt
# News Application – Python dependencies
# Install with: pip install -r requirements.txt

Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
Pillow>=10.0
requests>=2.31
flake8>=7.0
flake8-docstrings>=1.7
